Switch to backend check and make it a warning
