+++ tests/test_unittest/test_gtscript_frontend.py
@@ -605,44 +605,20 @@ class TestParallelComputations:
         module = f"TestParallelComputations_test_race_condition_{id_version}"
         externals = {}
 
-        with pytest.raises(
-            gt_frontend.GTScriptSyntaxError,
-            match="Horizontal race condition detected in stencil",
+        with pytest.warns(
+            Warning, match="Horizontal race condition detected in stencil",
         ):
 
+            @gtscript.stencil(backend="debug")
             def func_err(in_field: gtscript.Field[np.float_]):
                 with computation(PARALLEL), interval(...):
                     tmp = in_field
                     in_field = tmp[1, 0, 0]
 
-            compile_definition(func_err, "func_err", module)
-
+        @gtscript.stencil(backend="debug")
         def func(in_field: gtscript.Field[np.float_]):
             with computation(FORWARD), interval(...):
                 tmp = in_field
                 in_field = tmp[1, 0, 0]
 
-        compile_definition(func, "func", module)
-
-    def test_vertical_race_condition(self):
-        module = f"TestParallelComputations_test_race_condition_{id_version}"
-        externals = {}
-
-        with pytest.raises(
-            gt_frontend.GTScriptSyntaxError,
-            match="Vertical race condition detected in stencil",
-        ):
-
-            def func_err(in_field: gtscript.Field[np.float_]):
-                with computation(PARALLEL), interval(...):
-                    tmp = in_field
-                    in_field = tmp[0, 0, 1]
-
-            compile_definition(func_err, "func_err", module)
-
-        def func(in_field: gtscript.Field[np.float_]):
-            with computation(FORWARD), interval(...):
-                tmp = in_field
-                in_field = tmp[0, 0, 1]
-
-        compile_definition(func, "func", module)
+        # The vertical race condition is already fixed.
